updated dmls17, bumpped up max iteration

+++ dmls_gen/src/dmls_gen.cpp
@@ -17,7 +17,7 @@ void DmlsGen::set_word_visited(uint32_t word) {
     _word_visit_ids[invert_bits(rword, _word_length)] = _visit_id;
 }
 
-void DmlsGen::generate_dmls(std::vector<bool>& sequence, uint8_t word_length, uint32_t iterations,
+void DmlsGen::generate_dmls(std::vector<bool>& sequence, uint8_t word_length, uint64_t iterations,
         std::function<void(std::vector<bool>& sequence, uint8_t word_length)> new_record_callback) {
     if (word_length > 32) {
         return;
